+++ packages/stack-server/src/stack.tsx
@@ -6,7 +6,7 @@ if (process.env.NEXT_PUBLIC_STACK_PROJECT_ID !== "internal") {
   throw new Error("This project is not configured correctly. stack-server must always use the internal project.");
 }
 
-export const stackServerApp = new StackServerApp({
+export const stackServerApp = new StackServerApp<"nextjs-cookie", true, 'internal'>({
   tokenStore: "nextjs-cookie",
   urls: {
     afterSignIn: "/projects",