+++ packages/stack/src/components/redirect-message-card.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/navigation";
 import { useStackApp } from "..";
 import MessageCard from "./message-card";
 import { Text, Button } from "../components-core";
-import { neverResolve } from "@stackframe/stack-shared/dist/utils/promises";
+import { neverResolve, wait } from "@stackframe/stack-shared/dist/utils/promises";
 
 export default function RedirectMessageCard({ 
   type,
@@ -17,44 +17,42 @@ export default function RedirectMessageCard({
   const router = useRouter();
 
   let title: string;
-  let primaryUrl: string;
-  let secondaryUrl: string | null = null;
+  let primaryAction: () => Promise<void>;
   let message: string | null = null;
   let primaryButton: string;
   let secondaryButton: string | null = null;
   switch (type) {
     case 'signedIn': {
       title = "You are already signed in";
-      primaryUrl = stackApp.urls.home;
-      secondaryUrl = stackApp.urls.signOut;
+      primaryAction = () => stackApp.redirectToAfterSignOut();
       primaryButton = "Go to Home";
       secondaryButton = "Sign Out";
       break;
     }
     case 'signedOut': {
       title = "You are not currently signed in.";
-      primaryUrl = stackApp.urls.home;
+      primaryAction = () => stackApp.redirectToSignIn();
       primaryButton = "Go to Home";
       break;
     }
     case 'emailSent': {
       title = "Email sent!";
       message = 'Please check your inbox. Make sure to check your spam folder.';
-      primaryUrl = stackApp.urls.home;
+      primaryAction = () => neverResolve();
       primaryButton = "Go to Home";
       break;
     }
     case 'passwordReset': {
       title = "Password reset successfully!";
       message = 'Your password has been reset. You can now sign in with your new password.';
-      primaryUrl = stackApp.urls.signIn;
+      primaryAction = () => stackApp.redirectToSignIn();
       primaryButton = "Go to Sign In";
       break;
     }
     case 'emailVerified': {
       title = "Email verified!";
       message = 'Your have successfully verified your email.';
-      primaryUrl = stackApp.urls.home;
+      primaryAction = () => stackApp.redirectToSignIn();
       primaryButton = "Go to Home";
       break;
     }
@@ -69,18 +67,14 @@ export default function RedirectMessageCard({
           <Button
             variant="secondary"
             onClick={async () => {
-              router.push(stackApp.urls.signOut.toString());
-              await neverResolve();
+              await stackApp.redirectToSignOut();
             }}
           >
             {secondaryButton}
           </Button>
         )}
         
-        <Button onClick={async () => {
-          router.push(primaryUrl.toString());
-          await neverResolve();
-        }}>
+        <Button onClick={primaryAction}>
           {primaryButton}
         </Button>
       </div>