+++ packages/stack/src/utils/browser-script.tsx
@@ -1,19 +1,28 @@
 const script = () => {
+  const attributes = ['data-joy-color-scheme', 'data-mui-color-scheme', 'data-theme', 'data-color-scheme', 'class'];
+
   const observer = new MutationObserver((mutations) => {
     mutations.forEach((mutation) => {
-      if (mutation.attributeName === 'data-joy-color-scheme') {
-        const joyColorScheme = document.documentElement.getAttribute('data-joy-color-scheme');
-        if (!joyColorScheme) {
-          return;
+      for (const attributeName of attributes) {
+        if (mutation.attributeName === attributeName) {
+          const colorTheme = document.documentElement.getAttribute(attributeName);
+          if (!colorTheme) {
+            continue;
+          }
+          const darkMode = colorTheme.includes('dark');
+          const lightMode = colorTheme.includes('light');
+          if (!darkMode && !lightMode) {
+            continue;
+          }
+          document.documentElement.setAttribute('data-stack-theme', darkMode ? 'dark' : 'light');
         }
-        document.documentElement.setAttribute('data-theme', joyColorScheme);
       }
     });
   });
 
   observer.observe(document.documentElement, {
     attributes: true,
-    attributeFilter: ['data-joy-color-scheme'],
+    attributeFilter: attributes,
   });
 };