+++ apps/cjs-test/CHANGELOG.md
@@ -1,5 +1,27 @@
 # cjs-test
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.4.3
+
+## 0.2.2
+
+### Patch Changes
+
+- @stackframe package namespace
+- Updated dependencies
+  - @stackframe/stack@2.4.2
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.4.1
+
 ## 0.2.0
 
 ### Minor Changes