+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,31 @@
 # demo-app
 
+## 1.1.3
+
+### Patch Changes
+
+- New projects page
+- Updated dependencies
+  - @stackframe/stack-shared@2.4.2
+  - @stackframe/stack@2.4.3
+
+## 1.1.2
+
+### Patch Changes
+
+- @stackframe package namespace
+- Updated dependencies
+  - @stackframe/stack@2.4.2
+
+## 1.1.1
+
+### Patch Changes
+
+- Teams, permissions and RBAC
+- Updated dependencies
+  - @stackframe/stack-shared@2.4.1
+  - @stackframe/stack@2.4.1
+
 ## 1.1.0
 
 ### Minor Changes